move unit controller to subdir
