+++ api/src/controllers/units/UploadUnitController.ts
@@ -1,11 +1,8 @@
-import {Body, Put, Post, Delete, Param, JsonController, UseBefore, UploadedFile} from 'routing-controllers';
+import {Body, Post, JsonController, UseBefore, UploadedFile} from 'routing-controllers';
 import fs = require('fs');
-import passportJwtMiddleware from '../security/passportJwtMiddleware';
+import passportJwtMiddleware from '../../security/passportJwtMiddleware';
 
-import {Lecture} from '../models/Lecture';
-import {Unit} from '../models/units/Unit';
-import {IUnit} from '../../../shared/models/units/IUnit';
-import {VideoUnit} from '../models/units/VideoUnit';
+import {VideoUnit} from '../../models/units/VideoUnit';
 import {UnitController} from './UnitController';
 
 const uploadOptions = {dest: 'uploads/' };
